LayoutTests/imported/w3c/web-platform-tests/mediacapture-streams/GUM-echoCancellation-remote-only.https-expected.txt

When prompted, accept to share your audio stream.


FAIL getUserMedia suports "remote-only" assert_equals: expected (string) "remote-only" but got (boolean) true
